python/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
celery[redis]>=5.4.0
redis>=5.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
certifi>=2024.0.0
truststore>=0.9.0
python-dotenv>=1.0.0
pydantic-settings>=2.2.0
alembic>=1.13.0
Pillow>=10.0.0
